+++ Application/Application.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
     <UserSecretsId>aspnet-Dictionary_MVC-E85066DA-E827-4354-8577-B3A1A8446818</UserSecretsId>
-    <RootNamespace>Dictionary_MVC</RootNamespace>
+    <RootNamespace>Application</RootNamespace>
   </PropertyGroup>
